Refactor tests and helpers
