+++ tests/acs/test_acs_kernels.py
@@ -3,10 +3,10 @@ import os
 from stsci.tools import teal
 import drizzlepac
 from drizzlepac import astrodrizzle
-from drizzlepac.helpers.mark import require_bigdata
-from drizzlepac.helpers.io import get_bigdata
+from ..helpers.mark import require_bigdata
+from ..helpers.io import get_bigdata
 
-from ..helpers import BaseACS, raw_from_asn
+from ..resources import BaseACS, raw_from_asn
 
 def pytest_generate_tests(metafunc):
     # called once per each test function
