+++ tests/acs/test_asn_regress.py
@@ -3,15 +3,15 @@ import os
 from stsci.tools import teal
 import drizzlepac
 from drizzlepac import astrodrizzle
-from drizzlepac.helpers.mark import require_bigdata
+from ..helpers.mark import require_bigdata
 
-from ..helpers import BaseACS, raw_from_asn
+from ..resources import BaseACS, raw_from_asn
 
 
-class TestAsnNarrowband(BaseACS):
+class TestAsnRegress(BaseACS):
     
-    def test_acs_narrowband(self):
-        rootname = 'j8dw01010'
+    def test_hrc_asn(self):
+        rootname = 'j8bt06010'
         asn_file = rootname + '_asn.fits'
 
         # Prepare input files.
@@ -26,18 +26,20 @@ class TestAsnNarrowband(BaseACS):
         parObj['runfile'] = 'drizzlepac.run'
         parObj['STATE OF INPUT FILES']['preserve'] = False
         parObj['STATE OF INPUT FILES']['clean'] = True
-        parObj['STEP 2: SKY SUBTRACTION']['use_static'] = False
-        parObj['STEP 2: SKY SUBTRACTION']['sky_bits'] = None
+        parObj['STEP 1: STATIC MASK']['static_sig'] = 3.0
+        parObj['STEP 2: SKY SUBTRACTION']['skywidth'] = 0.1
+        parObj['STEP 2: SKY SUBTRACTION']['skylower'] = -50.0
+        parObj['STEP 2: SKY SUBTRACTION']['skyupper'] = 200.0
+        parObj['STEP 3: DRIZZLE SEPARATE IMAGES']['driz_sep_bits'] = 8578
         parObj['STEP 4: CREATE MEDIAN IMAGE']['combine_maskpt'] = 0.7
         parObj['STEP 4: CREATE MEDIAN IMAGE']['combine_nsigma'] = '6 3'
         parObj['STEP 4: CREATE MEDIAN IMAGE']['combine_nhigh'] = 1
         parObj['STEP 6: REMOVE COSMIC RAYS WITH DERIV, DRIZ_CR']['driz_cr_snr'] = '3.0 2.5'
         parObj['STEP 7: DRIZZLE FINAL COMBINED IMAGE']['final_bits'] = 8578
-        parObj['STEP 7a: CUSTOM WCS FOR FINAL OUTPUT']['final_wcs'] = True
-        parObj['STEP 7a: CUSTOM WCS FOR FINAL OUTPUT']['final_rot'] = 0.0
+        parObj['STEP 7: DRIZZLE FINAL COMBINED IMAGE']['final_units'] = 'counts'
         
         astrodrizzle.AstroDrizzle(asn_file, configobj=parObj)
 
         # Compare results
-        outputs = [('j8dw01010_drz.fits', 'reference_narrowband.fits')]
+        outputs = [('j8bt06011_drz.fits', 'reference_asn_regress.fits')]
         self.compare_outputs(outputs)
